fix(acp): remove bridge.rs, drop LocalSet from runner, fix Send bounds
- Delete transport/bridge.rs; inline spawn_agent_connection into http.rs
  using std::thread::spawn + current_thread runtime + LocalSet so !Send
  agent futures run correctly per connection
- Remove AgentSpawner Send bound on future (Agent<LoopbackChannel> is
  !Send by design); switch all session-spawn call sites to spawn_local
- Replace LocalSet::new().run_until() in runner.rs AcpTransport::Both
  branch with direct tokio::select! on both server futures
- Gate integration.rs (ACP 0.10 API) with #[cfg(any())] until PR 3
- Fix ProtocolVersion path in transport/tests.rs

BREAKING CHANGE: AgentSpawner future no longer requires + Send;
transport::bridge module removed, use transport::http::spawn_agent_connection
