+++ crates/zeph-acp/src/transport/mod.rs
@@ -20,8 +20,6 @@
 #[cfg(feature = "acp-http")]
 pub mod auth;
 #[cfg(feature = "acp-http")]
-pub mod bridge;
-#[cfg(feature = "acp-http")]
 pub mod discovery;
 pub mod http;
 pub mod router;
